requirements.txt
# ============================================================
# AI RC Car — Python Dependencies
# ============================================================

# Core
paho-mqtt>=2.0.0
requests>=2.28.0
colorama>=0.4.6
customtkinter>=5.2.0

# AI Assistant — Text-to-Speech
pyttsx3>=2.90

# Voice Control
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

# Web Dashboard
flask>=3.0.0
flask-socketio>=5.3.0

# Drowsiness Detection
opencv-python>=4.8.0
mediapipe==0.10.14

# JSON (standard library — no install needed)
# ArduinoJson is an ESP32/Arduino library, install via Arduino IDE
# Blynk library — install via Arduino IDE Library Manager
